reorganized chromosome generation to the train_chrom step
