+++ genetic_algorithm_vae_predictor_Multi_Computer.py
@@ -26,230 +26,6 @@ from vaelstmpredictor.vae_predictor.train import train_vae_predictor
 
 from GeneticAlgorithm import *
 
-from paramiko import SSHClient, SFTPClient, Transport, AutoAddPolicy, ECDSAKey
-
-import multiprocessing as mp
-
-def train_generation(generation, clargs, private_key='id_ecdsa'):
-	getChrom = 'https://philippesaade11.pythonanywhere.com/GetChrom'
-	getFitness = 'http://philippesaade11.pythonanywhere.com/GetFitness'
-
-	private_key = os.environ['HOME'] + '/.ssh/{}'.format(private_key)
-	key_filename = private_key
-	
-	machines = [# {"host": "192.168.0.1", "username": "acc", 
-				#   "key_filename": key_filename},
-				{"host": "172.16.50.181", "username": "acc", 
-					"key_filename": key_filename},
-				# {"host": "172.16.50.176", "username": "acc", 
-					# "key_filename": key_filename},
-				{"host": "172.16.50.177", "username": "acc", 
-					"key_filename": key_filename},
-				{"host": "172.16.50.163", "username": "acc", 
-					"key_filename": key_filename},
-				{"host": "172.16.50.182", "username": "acc", 
-					"key_filename": key_filename},
-				{"host": "172.16.50.218", "username": "acc", 
-					"key_filename": key_filename},
-				{"host": "172.16.50.159", "username": "acc", 
-					"key_filename": key_filename},
-				{"host": "172.16.50.235", "username": "acc", 
-					"key_filename": key_filename},
-				{"host": "172.16.50.157", "username": "acc", 
-					"key_filename": key_filename},
-				{"host": "172.16.50.237", "username": "acc", 
-					"key_filename": key_filename}
-				]
-	
-	queue = mp.Queue()
-	
-	#Create Processes
-	for machine in machines: queue.put(machine)
-	
-	alldone = False
-	while not alldone:
-		alldone = True
-		for chrom in generation:
-			if not chrom.isTrained:
-				print("Creating Process for Chromosome {}".format(
-							chrom.chromosomeID), end=" on machine ")
-				# Find a Chromosome that is not trained yet
-				alldone = False
-				
-				# Wait for queue to have a value, 
-				#	which is the ID of the machine that is done.
-				machine = queue.get()
-				print('{}'.format(machine['host']))
-				process = mp.Process(target=train_chromosome, 
-									args=(chrom, machine, queue))
-				process.start()
-				
-				clargs.generationID = chrom.generationID
-				clargs.chromosomeID = chrom.chromosomeID
-
-				table_dir = clargs.table_dir
-				table_name = '{}/{}_{}_{}_sql_fitness_table_{}.json'
-				table_name= table_name.format(clargs.table_dir, 
-										clargs.run_name, clargs.generationID, 
-										clargs.chromosomeID, clargs.time_stamp)
-
-				json_ID = {'generationID':clargs.generationID,
-						   'chromosomeID':clargs.chromosomeID}
-
-				sql_json = requests.get(getFitness, params=json_ID).json()
-
-				if isinstance(sql_json, dict):
-					with open(table_name, 'a') as f_out:
-					    json.dump(sql_json, f_out)
-					
-					chrom.fitness = sql_json['fitness']
-				else:
-					print('SQL Request Failed: sql_json = {}'.format(sql_json))
-					chrom.fitness = sql_json or 0
-
-def train_chromosome(chromosome, machine, queue, port=22, logdir='train_logs',
-					 zip_filename="vaelstmpredictor.zip", verbose=True):
-	
-	if not os.path.exists(logdir): os.mkdir('train_logs')
-	if verbose: print('[INFO] Checking if file {} exists on {}'.format(zip_filename, machine['host']))
-
-	# sys.stdout = open(logdir+'/output'+str(chromosome.chromosomeID)+".txt", 'w')
-	# sys.stderr = open(logdir+'/error'+str(chromosome.chromosomeID)+".txt", 'w')
-	
-	
-	ssh = SSHClient()
-	ssh.set_missing_host_key_policy(AutoAddPolicy())
-	ssh.connect(machine["host"], key_filename=machine['key_filename'])
-
-	stdin, stdout, stderr = ssh.exec_command('ls | grep vaelstmpredictor')
-	
-	if(len(stdout.readlines()) == 0):
-		if verbose: 
-			print('[INFO] File {} does not exists on {}'.format(
-									zip_filename, machine['host']))
-		
-		#Upload Files to Machine
-		print("Uploading file to machine")
-		
-		if verbose: print('[INFO] Transfering {} to {}'.format(zip_filename, machine['host']))
-		transport = Transport((machine["host"], port))
-		pk = ECDSAKey.from_private_key(open(machine['key_filename']))
-		transport.connect(username = machine["username"], pkey=pk)
-		
-		sftp = SFTPClient.from_transport(transport)
-		sftp.put(zip_filename, zip_filename)
-		
-		stdin, stdout, stderr = ssh.exec_command('unzip {}'.format(zip_filename))
-		error = "".join(stderr.readlines())
-		if error != "":
-			print("Errors has occured while unzipping file in machine: "+str(machine)+"\nError: "+error)
-				
-		stdin, stdout, stderr = ssh.exec_command('cd vaelstmpredictor; '
-						'../anaconda3/envs/tf_gpu/bin/python setup.py install')
-		error = "".join(stderr.readlines())
-		if error != "":
-			print("Errors setting up vaelstmpredictor: "
-					+ str(machine)
-					+ "\nError: "
-					+ error)
-
-		sftp.close()
-		transport.close()
-		print("File uploaded")
-	else:
-		if verbose: 
-			print('[INFO] File {} exists on {}'.format(zip_filename, 
-													machine['host']))
-
-	# transport = Transport((machine["host"], port))
-	# pk = ECDSAKey.from_private_key(open(machine['key_filename']))
-	# transport.connect(username = machine["username"], pkey=pk)
-	
-	# sftp = SFTPClient.from_transport(transport)
-	# sftp.put(param_filename, 'vaelstmpredictor/{}'.format(param_filename))
-	# sftp.close()
-	# transport.close()
-	
-	command = []
-	command.append('cd vaelstmpredictor; ')
-	command.append('../anaconda3/envs/tf_gpu/bin/python run_chromosome.py ')
-	command.append('--table_dir {} '.format(clargs.table_dir))
-	command.append('--generationID {} '.format(chromosome.generationID))
-	command.append('--chromosomeID {} '.format(chromosome.chromosomeID))
-	print('[DEBUG] 1 command: {}'.format(command))
-	for key,val in clargs.__dict__.items():
-		if key not in ['generationID', 'chromosomeID']:
-			command.append('--{} {}'.format(key,val))
-	
-	print('[DEBUG] 2 command: {}'.format(command))
-	command.append('--num_vae_layers {} '.format(chromosome.num_vae_layers))
-	command.append('--num_dnn_layers {} '.format(chromosome.num_dnn_layers))
-	command.append('--size_vae_latent {} '.format(chromosome.size_vae_latent))
-	command.append('--size_vae_hidden {} '.format(chromosome.size_vae_hidden))
-	command.append('--size_dnn_hidden {} '.format(chromosome.size_dnn_hidden))
-	
-	print('[DEBUG] 3 command: {}'.format(command))
-	command = " ".join(command)
-	
-	print("Executing command:\n\t{}".format(command))
-	
-	stdin, stdout, stderr = ssh.exec_command(command)
-	
-	try:
-		stdout.channel.recv_exit_status()
-		for line in stdout.readlines(): print(line)
-	except Exception as e:
-		print('error on stdout.readlines(): {}'.format(str(e)))
-
-	try:
-		stderr.channel.recv_exit_status()
-		for line in stderr.readlines(): print(line)
-	except Exception as e:
-		print('error on stderr.readlines(): {}'.format(str(e)))
-
-	try:
-		stdin.channel.recv_exit_status()
-		for line in stdin.readlines(): print(line)
-	except Exception as e:
-		print('error on stdin.readlines(): {}'.format(str(e)))
-
-	error = "".join(stderr.readlines())
-	if error != "":
-		print("Errors has occured while tainging in machine: "+str(machine)+"\nError: "+error)
-	if "".join(stdout.readlines()[-4:]) == "done":
-		print("Trained Successfully")
-
-	# transport = Transport((machine["host"], port))
-	# pk = ECDSAKey.from_private_key(open(machine['key_filename']))
-	# transport.connect(username = machine["username"], pkey=pk)
-	# 
-	# sftp = SFTPClient.from_transport(transport)
-	# sftp.put(param_filename, 'vaelstmpredictor/{}'.format(param_filename))
-	# 
-	# sftp.close()
-	# transport.close()
-
-	queue.put(machine)
-
-	table_dir = clargs.table_dir
-	table_name = '{}/{}_{}_{}_fitness_table_{}.csv'
-	table_name = table_name.format(clargs.table_dir, 
-								clargs.run_name, clargs.generationID, 
-								clargs.chromosomeID, clargs.time_stamp)
-
-	with open(table_name, 'r') as f_in:
-		check = 'fitness:'
-		for line in f_in.readlines():
-			if 'generationID:{}'.format(clargs.generationID) in line:
-				if 'chromosomeID:{}'.format(clargs.chromosomeID) in line:
-					fitness = line.split(check)[1].split(',')[0]
-					chromosome.fitness = float(fitness)
-					break
-
-	chromosome.isTrained = True
-	print("Command Executed")
-	ssh.close()				
-
 if __name__ == '__main__':
 	parser = argparse.ArgumentParser()
 	parser.add_argument('run_name', type=str, default='ga_test_',
@@ -357,12 +133,14 @@ if __name__ == '__main__':
 
 	clargs.n_labels = len(np.unique(data_instance.train_labels))
 
-	generation = generate_random_chromosomes(population_size = population_size,
-											 clargs = clargs, 
-											 data_instance = data_instance, 
-											 TrainFunction = train_generation)
+	generation = generate_random_chromosomes(population_size = population_size)
+											 # clargs = clargs, 
+											 # data_instance = data_instance, 
+											 # TrainFunction = train_generation
 
-	generationID = 0	
+	generation = train_generation(generation, clargs)
+
+	generationID = 0
 	evolutionary_tree = {}
 	evolutionary_tree[generationID] = save_generation_to_tree(generation,
 															verbose=verbose)
